refactor: redis에 리포트 생성 job의 상태를 저장하는 기능 에러 수정.
